backend/requirements.txt
fastapi[standard]
uvicorn[standard]==0.30.6
pymongo==4.10.1
python-multipart==0.0.9
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
certifi==2026.2.27
